refactor available hours endpoint Co-authored-by: Greg Magkos <grigmag@users.noreply.github.com>
